scripts/requirements.txt
wheel
pandas
bs4
lxml
textract
polyglot
pycld2
PyICU
normality
delegator.py
git+git://github.com/occrp/countrynames.git
